Use UInt instead Int where possible; some fixes in hashing method

+++ src/main/kotlin/IndexBuilder.kt
@@ -5,20 +5,23 @@ import java.nio.ByteBuffer
  * Builder for a binary searchable file.
  */
 class IndexBuilder() {
-    private val list = mutableListOf<Long>()
+    private val list = mutableListOf<ULong>()
 
-    fun add(hash: Int, id: Int) {
-        val entry = (hash.toLong() shl 32) or id.toLong()
+    fun add(hash: UInt, id: UInt) {
+        val entry = (hash.toULong() shl 32) or id.toULong()
         list.add(entry)
     }
 
     fun write(filename: String) {
+        // sort and build buffer
         list.sort()
         val buffer = ByteBuffer.allocate(list.size * 8)
         val longBuffer = buffer.asLongBuffer()
         for(entry in list) {
-            longBuffer.put(entry)
+            longBuffer.put(entry.toLong())
         }
+
+        // write buffer
         FileOutputStream(filename).use { fos ->
             fos.channel.write(buffer)
         }
